REFACTOR: Server binding to PORT on a different file, allow for app export for testing purposes
