+++ package.json
@@ -21,5 +21,9 @@
     "mongoose": "^6.3.6",
     "morgan": "^1.10.0",
     "socket.io": "^4.5.1"
+  },
+  "devDependencies": {
+    "jest": "^28.1.3",
+    "supertest": "^6.2.4"
   }
 }
